Shapes have neighbours
Shapes index in a List of Voronoi shapes is the same as their DelPoint. Therefore they can contain a list of ints that can reference their neighbour shape from the List they exist in.

DelPoints and DelTriangles were changed to classes as it was easier to make them reference types for variable editing purposes.