Various Toys and Changes
Made many toys to be able to explore Voronoi patterns and the problems I was having.
Seemed to mostly solve the box clamp of shapes, but there still remains a few niche cases where the shapes do not get clamped.
One of these problems seems to be a float not comparing with another float correctly.